feat: Bring back some commands, install everything to /usr/5bin etc
I had an idea of, instead of stripping down Heirloom and letting some historical tools out, letting it in its completeness and linking some of its commands that are not supplied by lobase linked on /usr/bin and /bin. /sbin will not be conflicting with nothing, so we shall not worry.

This is, by far, the biggest change in Copacabana's design. I do not intend to change nothing more.